// src/app/interfaces/Firmas_Terceros.interface.ts
export interface firmaTercero {
    id_cliente:             string;
    id_afiliado_estado:     number;
    id_parentesco:          number;
    id_firma:               number;
    codigo_afiliado:        string;
    cuenta_afiliado:        number;
    id_filial_pertenece:    number;
    apertura_actualizacion: string;
    id_colaborador:         number;
    observaciones:          string;
}